fix: 对齐官方长消息分段基线并迁移 markdown_strip
- 默认 QQ_*_CHUNK_SOFT_LIMIT 调整为 5000，对齐官方非流式长消息分段基线
- 明确普通分段发送不主动 sleep，只逐段 await 成功后立即发送下一段
- 将 markdown_strip 实现迁移到 qq-maid-common，core 保留兼容入口
