+++ runtime/config/.env.example
@@ -16,9 +16,9 @@ QQ_MAID_ENABLE_IMAGE=false
 
 # 普通回复分段软限制（字符数）。这是 Gateway 保守软上限，并非 QQ 平台已确认的硬上限。
 # 超过软限制的普通回复会被拆成多条按 Markdown 安全边界发送，不再被 Core 静默截断。
-# 真实平台限制仍需后续真机验证。
-QQ_MARKDOWN_CHUNK_SOFT_LIMIT=1800
-QQ_TEXT_CHUNK_SOFT_LIMIT=1800
+# 默认对齐官方非流式长消息分段的 5000 字符基线；真实平台限制仍需后续真机验证。
+QQ_MARKDOWN_CHUNK_SOFT_LIMIT=5000
+QQ_TEXT_CHUNK_SOFT_LIMIT=5000
 
 # 群消息处理策略：
 # off=忽略，command=只处理 / 命令，mention=命令/@/回复机器人，active=处理包含指定提示词的普通群消息。
